refactor(rate-service): add usage of proto generated types and delete constructor of abstract rate client
