+++ rate-service/nest-cli.json
@@ -1,7 +1,7 @@
 {
   "$schema": "https://json.schemastore.org/nest-cli",
   "collection": "@nestjs/schematics",
-  "sourceRoot": "src",
+  "sourceRoot": "rate-service/src",
   "compilerOptions": {
     "deleteOutDir": true,
     "assets": [
