+++ rate-service/package.json
@@ -22,10 +22,12 @@
   "dependencies": {
     "@grpc/grpc-js": "^1.10.10",
     "@grpc/proto-loader": "^0.7.13",
+    "@nestjs/axios": "^3.0.2",
     "@nestjs/common": "^10.0.0",
     "@nestjs/core": "^10.0.0",
     "@nestjs/microservices": "^10.3.10",
     "@nestjs/platform-express": "^10.0.0",
+    "axios": "^1.7.2",
     "reflect-metadata": "^0.2.0",
     "rxjs": "^7.8.1"
   },
