+++ rate-service/src/app.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 import { RateModule } from './rate/rate.module';
-import { AppConfigModule } from './config/app-config.module';
+import { AppConfigModule } from './config/app-config';
 import { LoggerModule } from 'nestjs-pino';
 
 @Module({
