+++ rate-service/src/config/app-config/app-config.service.ts
