+++ rate-service/src/config/app-config/index.ts
@@ -1,0 +1,2 @@
+export * from './app-config.module';
+export * from './app-config.service';
