+++ rate-service/src/main.ts
@@ -2,6 +2,7 @@ import { NestFactory } from '@nestjs/core';
 import { AppModule } from './app.module';
 import { MicroserviceOptions, Transport } from '@nestjs/microservices';
 import { join } from 'path';
+import { RATE_PACKAGE_NAME } from '../../proto/dist/types/rate';
 
 async function bootstrap() {
   const app = await NestFactory.createMicroservice<MicroserviceOptions>(
@@ -9,9 +10,9 @@ async function bootstrap() {
     {
       transport: Transport.GRPC,
       options: {
-        package: 'rate',
-        protoPath: join(__dirname, '..', '..', 'proto/rate/rate.proto'),
-        url: '0.0.0.0:5001',
+        package: RATE_PACKAGE_NAME,
+        protoPath: join(__dirname, '../../../../proto/rate/rate.proto'),
+        url: `${process.env.HOST}:${process.env.PORT}`,
       },
     },
   );
