+++ rate-service/src/rate/clients/abstract-rate.client.ts
@@ -1,20 +1,9 @@
-import { HttpService } from '@nestjs/axios';
-import { InjectPinoLogger, PinoLogger } from 'nestjs-pino';
-
-import { AppConfigService } from '../../config/app-config.service';
 import { RateClientException } from '../exceptions';
 import { IExchangeRate } from '../interfaces';
 import { IRateClient } from '../interfaces/rate-client.interface';
 
 export abstract class AbstractRateClient implements IRateClient {
   private next: IRateClient;
-
-  constructor(
-    protected readonly httpService: HttpService,
-    protected readonly appConfigService: AppConfigService,
-    @InjectPinoLogger()
-    protected readonly logger: PinoLogger,
-  ) {}
 
   setNext(client: IRateClient): IRateClient {
     this.next = client;
