+++ rate-service/src/rate/rate.constants.ts
@@ -2,4 +2,4 @@ export const SECOND = 1000;
 
 export const DOLLAR_ABBREVIATION = 'USD';
 
-export const DOLLAR_CODE = 840;
+export const MONOBANK_DOLLAR_CODE = 840;
